Rename package from 'pgcrypto_fields' to 'pgcrypto'
