+++ .coveragerc
@@ -1,3 +1,3 @@
 [run]
-source = pgcrypto_fields
+source = pgcrypto
 omit = *migrations*, *tests*
