+++ pgcrypto/aggregates.py
