+++ setup.cfg
@@ -4,5 +4,5 @@ max-complexity = 10
 exclude = *migrations*
 ignore = D100,D203,D204
 statistics=true
-application-import-names = pgcrypto_fields
+application-import-names = pgcrypto
 import-order-style = google
